fix(*): update ts and typedoc dependencies
refactor handlebars helpers registration and leave only dist folder as
package distribution

Closes #6
